[MOD/#11] HomeFragment ->FriendFragment
