+++ app/src/main/java/org/sopt/dosopttemplate/adapter/FriendBirthdayViewHolder.kt
@@ -2,7 +2,7 @@ package org.sopt.dosopttemplate.adapter
 
 import androidx.recyclerview.widget.RecyclerView
 import org.sopt.dosopttemplate.databinding.ItemFriendBirthdayBinding
-import org.sopt.dosopttemplate.repository.model.data.Friend
+import org.sopt.dosopttemplate.model.data.Friend
 
 class FriendBirthdayViewHolder(private val binding: ItemFriendBirthdayBinding) :
     RecyclerView.ViewHolder(binding.root) {
