+++ app/src/main/java/org/sopt/dosopttemplate/adapter/FriendMyViewHolder.kt
@@ -2,7 +2,7 @@ package org.sopt.dosopttemplate.adapter
 
 import androidx.recyclerview.widget.RecyclerView
 import org.sopt.dosopttemplate.databinding.ItemFriendMyBinding
-import org.sopt.dosopttemplate.repository.model.data.Friend
+import org.sopt.dosopttemplate.model.data.Friend
 
 class FriendMyViewHolder(private val binding: ItemFriendMyBinding) :
     RecyclerView.ViewHolder(binding.root) {
