+++ app/src/main/java/org/sopt/dosopttemplate/model/data/Friend.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.repository.model.data
+package org.sopt.dosopttemplate.model.data
 
 import androidx.annotation.DrawableRes
