+++ app/src/main/java/org/sopt/dosopttemplate/ui/mypage/MyPageFragment.kt
@@ -8,8 +8,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import org.sopt.dosopttemplate.databinding.FragmentMyPageBinding
-import org.sopt.dosopttemplate.repository.model.UserSharedPreferences
-import org.sopt.dosopttemplate.repository.model.data.UserInfo
+import org.sopt.dosopttemplate.model.UserSharedPreferences
+import org.sopt.dosopttemplate.model.data.UserInfo
 import org.sopt.dosopttemplate.ui.auth.LoginActivity
 import org.sopt.dosopttemplate.util.base.BindingFragment
