+++ app/src/main/res/menu/home_bnv.xml
@@ -9,6 +9,10 @@
 
     <!-- Home -->
     <item
+        android:id="@+id/item_friend"
+        android:icon="@drawable/ic_person_white_24"
+        android:title="친구" />
+    <item
         android:id="@+id/item_home"
         android:icon="@drawable/ic_home_white_24"
         android:title="홈" />
